Update list.tsv: Add DESC and DESC_JP columns (#3), Add U+8A01 (#4)
list.tsv is updated so that JP only contains reading pronunciation
of a character. All characters except the three below are monosyllabic.

```
𠯢	U+E064	saa1 aa6
〺	U+5345	saa1 aa6
卌	U+534C	sei3 aa6
```
Two columns, `DESC` and `DESC_JP` have been added to capture the
descriptive reading of a character, which fall into two catagories:
(a) characters for measurement and (b) radicals.
`DESC` is the description of the character.
`DESC_JP` is the Jyutping pronunciation of the description.
For these characters, `JP` will be empty.

See #3 for more details.

This entry was missing and it is now fixed. (#4)
```
訁	U+8A01					言字邊	jin4 zi6 bin1
```